+++ components/internal/StacFields.vue
@@ -118,9 +118,6 @@ export default {
 <style>
 @import url('../base.css');
 
-.vue-component.stac-fields .unit {
-	opacity: 0.6;
-}
 .vue-component.stac-fields .table {
 	width: 100%;
 	border-collapse: collapse;